Fix invalid `#valid_password?` to correctly validate a password
The bug being fixed was discovered with the new `test/integration/change_password_test.rb`.

Major changes to test setup and layout, precipitated by adding a second integration test spec file. The previous duplicated setup code in both integration test specs caused an error when the second spec was run, as `Hanami::Model::Configuration` raised a `Hamami::Model::Error` saying that the "Relation with `register_as :users` registered more than once".

That precipitated replacing the copy-pasta with `require` statements including files in `test/support`. Since Ruby evaluates `require`d files only once per invocation of Ruby (in this case, by `rake`), the duplicate registration never occurs.

72 tests, 86 assertions, 0 failures, 0 errors, 0 skips
Coverage: 987 / 987 LOC (100.0%) covered.
RuboCop: 11 files inspected, no offenses detected
Flay: Total score 0
Flog: Total 310.7; method average 4.5; max 9.5 (CryptIdent::SignIn#call)
Reek: 0 total warnings
Inch: Nothing to suggest
