+++ docs/_index.html
@@ -106,7 +106,7 @@
 </div>
 
       <div id="footer">
-  Generated on Sat Jan 26 22:36:53 2019 by
+  Generated on Tue Jan 29 03:20:09 2019 by
   <a href="http://yardoc.org" title="Yay! A Ruby Documentation Tool" target="_parent">yard</a>
   0.9.16 (ruby-2.6.0).
 </div>
